Move auth form to separated component
